+++ Venmo/src/main/java/com/braintreepayments/api/VenmoLauncherCallback.java
@@ -1,5 +1,7 @@
 package com.braintreepayments.api;
 
+import androidx.annotation.NonNull;
+
 /**
  * Used to receive notification that the Venmo payment authorization flow completed
  * Once this is invoked, continue the flow by calling
@@ -7,5 +9,5 @@ package com.braintreepayments.api;
  */
 public interface VenmoLauncherCallback {
 
-    void onVenmoResult(VenmoPaymentAuthResult venmoPaymentAuthResult);
+    void onVenmoPaymentAuthResult(@NonNull VenmoPaymentAuthResult venmoPaymentAuthResult);
 }